+++ src/lib/components/EvmMethodChip.tsx
@@ -5,10 +5,13 @@ import { getEvmMethod } from "lib/utils";
 
 interface EvmMethodChipProps {
   txInput: string;
-  width: TagProps["width"];
+  width?: TagProps["width"];
 }
 
-export const EvmMethodChip = ({ txInput, width }: EvmMethodChipProps) => (
+export const EvmMethodChip = ({
+  txInput,
+  width = "144px",
+}: EvmMethodChipProps) => (
   <Tag width={width} height="17px" variant="gray" justifyContent="center">
     {getEvmMethod(txInput)}
   </Tag>